move block request to tab-level, and set response as context for children
